+++ apps/app/misc/community/index.tsx
@@ -1,15 +1,9 @@
+import { AtomIcon, BarChartIcon, CheckMarkIcon, LineChartIcon, PieChartIcon } from "../../icons";
 import { ChevronRightIcon } from "@heroicons/react/20/solid";
 import { post } from "datagovmy-ui/api";
 import { Button, Card, Container, Dropdown, Input, Spinner, toast } from "datagovmy-ui/components";
 import { clx } from "datagovmy-ui/helpers";
 import { useData, useTranslation } from "datagovmy-ui/hooks";
-import {
-  AtomIcon,
-  BarChartIcon,
-  CheckMarkIcon,
-  LineChartIcon,
-  PieChartIcon,
-} from "datagovmy-ui/icons";
 import { OptionType } from "datagovmy-ui/types";
 import { FunctionComponent } from "react";
