+++ apps/app/misc/helpdesk/index.tsx
@@ -1,3 +1,4 @@
+import { CheckMarkIcon } from "../../icons";
 import { ChevronRightIcon } from "@heroicons/react/20/solid";
 import { post } from "datagovmy-ui/api";
 import {
@@ -13,7 +14,6 @@ import {
 } from "datagovmy-ui/components";
 import { clx } from "datagovmy-ui/helpers";
 import { useData, useTranslation } from "datagovmy-ui/hooks";
-import { CheckMarkIcon } from "datagovmy-ui/icons";
 import { OptionType } from "datagovmy-ui/types";
 import { FunctionComponent } from "react";
