+++ apps/opendosm/icons/index.tsx
@@ -1,0 +1,234 @@
+import { IconProps } from "datagovmy-ui/icons";
+import { FunctionComponent } from "react";
+
+export const UsersIcon: FunctionComponent<IconProps> = ({ className }) => {
+  return (
+    <svg
+      width="22"
+      height="22"
+      viewBox="0 0 22 22"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      className={className}
+    >
+      <g clipPath="url(#clip0_2886_13900)">
+        <path
+          d="M8.33355 12.5555C6.25355 12.5555 2.11133 13.5955 2.11133 15.6667V17.2222H14.5558V15.6667C14.5558 13.5955 10.4135 12.5555 8.33355 12.5555ZM4.19133 15.4444C4.93799 14.9289 6.74244 14.3333 8.33355 14.3333C9.92466 14.3333 11.7291 14.9289 12.4758 15.4444H4.19133ZM8.33355 11C10.0491 11 11.4447 9.60444 11.4447 7.88888C11.4447 6.17333 10.0491 4.77777 8.33355 4.77777C6.61799 4.77777 5.22244 6.17333 5.22244 7.88888C5.22244 9.60444 6.61799 11 8.33355 11ZM8.33355 6.55555C9.07133 6.55555 9.66688 7.1511 9.66688 7.88888C9.66688 8.62666 9.07133 9.22222 8.33355 9.22222C7.59577 9.22222 7.00022 8.62666 7.00022 7.88888C7.00022 7.1511 7.59577 6.55555 8.33355 6.55555ZM14.5913 12.6089C15.6224 13.3555 16.3335 14.3511 16.3335 15.6667V17.2222H19.8891V15.6667C19.8891 13.8711 16.778 12.8489 14.5913 12.6089ZM13.6669 11C15.3824 11 16.778 9.60444 16.778 7.88888C16.778 6.17333 15.3824 4.77777 13.6669 4.77777C13.1869 4.77777 12.7424 4.89333 12.3335 5.08888C12.8935 5.87999 13.2224 6.84888 13.2224 7.88888C13.2224 8.92888 12.8935 9.89777 12.3335 10.6889C12.7424 10.8844 13.1869 11 13.6669 11Z"
+          fill="currentColor"
+        />
+      </g>
+      <defs>
+        <clipPath id="clip0_2886_13900">
+          <rect
+            width="21.3333"
+            height="21.3333"
+            fill="white"
+            transform="translate(0.333496 0.333313)"
+          />
+        </clipPath>
+      </defs>
+    </svg>
+  );
+};
+
+export const EconomicGrowthIcon: FunctionComponent<IconProps> = ({ className }) => {
+  return (
+    <svg
+      width="19"
+      height="12"
+      viewBox="0 0 19 12"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      className={className}
+    >
+      <path
+        d="M13.0558 0.666626L15.0913 2.70218L10.7535 7.03996L7.19799 3.4844L0.611328 10.08L1.86466 11.3333L7.19799 5.99996L10.7535 9.55551L16.3535 3.9644L18.3891 5.99996V0.666626H13.0558Z"
+        fill="currentColor"
+      />
+    </svg>
+  );
+};
+
+export const BankIcon: FunctionComponent<IconProps> = ({ className }) => {
+  return (
+    <svg
+      width="17"
+      height="18"
+      viewBox="0 0 17 18"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      className={className}
+    >
+      <path
+        d="M4.11133 8.22223H2.33355V14.4445H4.11133V8.22223ZM9.44466 8.22223H7.66688V14.4445H9.44466V8.22223ZM17.0002 16.2222H0.111328V18H17.0002V16.2222ZM14.778 8.22223H13.0002V14.4445H14.778V8.22223ZM8.55577 2.23112L13.1869 4.66667H3.92466L8.55577 2.23112ZM8.55577 0.222229L0.111328 4.66667V6.44445H17.0002V4.66667L8.55577 0.222229Z"
+        fill="currentColor"
+      />
+    </svg>
+  );
+};
+
+export const UnemploymentIcon: FunctionComponent<IconProps> = ({ className }) => {
+  return (
+    <svg
+      width="19"
+      height="18"
+      viewBox="0 0 19 18"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      className={className}
+    >
+      <path
+        d="M11.8229 7.22219L9.50022 9.54488L7.17755 7.22219L5.94466 8.45509L8.26732 10.7778L5.94471 13.1004L7.1776 14.3333L9.50022 12.0107L11.8228 14.3333L13.0557 13.1004L10.7331 10.7778L13.0558 8.45509L11.8229 7.22219Z"
+        fill="currentColor"
+      />
+      <path
+        fillRule="evenodd"
+        clipRule="evenodd"
+        d="M18.3891 5.44442C18.3891 4.45775 17.598 3.66664 16.6113 3.66664H13.0558V1.88886C13.0558 0.902195 12.2647 0.111084 11.278 0.111084H7.72244C6.73577 0.111084 5.94466 0.902195 5.94466 1.88886V3.66664H2.38911C1.40244 3.66664 0.620217 4.45775 0.620217 5.44442L0.611328 16.1111C0.611328 17.0977 1.40244 17.8889 2.38911 17.8889H16.6113C17.598 17.8889 18.3891 17.0977 18.3891 16.1111V5.44442ZM11.278 1.88886V3.66664H7.72244V1.88886H11.278ZM2.38911 16.1111V5.44442H16.6113V16.1111H2.38911Z"
+        fill="currentColor"
+      />
+    </svg>
+  );
+};
+
+export const InflationIcon: FunctionComponent<IconProps> = ({ className }) => {
+  return (
+    <svg
+      width="18"
+      height="12"
+      viewBox="0 0 18 12"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      className={className}
+    >
+      <path
+        d="M1.44466 11.7689L6.77799 6.42663L10.3335 9.98219L17.8891 1.48441L16.6358 0.231079L10.3335 7.31552L6.77799 3.75997L0.111328 10.4355L1.44466 11.7689Z"
+        fill="currentColor"
+      />
+    </svg>
+  );
+};
+
+export const ProductionIcon: FunctionComponent<IconProps> = ({ className }) => {
+  return (
+    <svg
+      width="19"
+      height="18"
+      viewBox="0 0 19 18"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      className={className}
+    >
+      <path
+        d="M18.3891 17.8889H0.611328V7.22219L6.83355 4.55553V6.33331L11.278 4.55553V7.22219H13.9447L14.8335 0.111084H17.5002L18.3891 7.22219V17.8889ZM9.50022 7.17775L5.05577 8.95553V7.22219L2.38911 8.39553V16.1111H16.6113V8.99997H9.50022V7.17775ZM8.61133 14.3333H10.3891V10.7778H8.61133V14.3333ZM5.05577 14.3333H6.83355V10.7778H5.05577V14.3333ZM13.9447 10.7778H12.1669V14.3333H13.9447V10.7778Z"
+        fill="currentColor"
+      />
+    </svg>
+  );
+};
+
+export const IndustryIcon: FunctionComponent<IconProps> = ({ className }) => {
+  return (
+    <svg
+      width="16"
+      height="16"
+      viewBox="0 0 16 16"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      className={className}
+    >
+      <path
+        d="M5.3335 8.93336C5.3335 10.3822 6.52461 11.5556 8.00016 11.5556C9.47572 11.5556 10.6668 10.3822 10.6668 8.93336C10.6668 7.76892 10.1957 7.43114 8.00016 4.88892C5.79572 7.44892 5.3335 7.77781 5.3335 8.93336Z"
+        fill="currentColor"
+      />
+      <path
+        d="M15.1111 8.88889C15.6 8.88889 16 8.48889 16 8C16 7.51111 15.6 7.11111 15.1111 7.11111H14.2222V1.77778H15.1111C15.6 1.77778 16 1.37778 16 0.888889C16 0.4 15.6 0 15.1111 0H0.888889C0.4 0 0 0.4 0 0.888889C0 1.37778 0.4 1.77778 0.888889 1.77778H1.77778V7.11111H0.888889C0.4 7.11111 0 7.51111 0 8C0 8.48889 0.4 8.88889 0.888889 8.88889H1.77778V14.2222H0.888889C0.4 14.2222 0 14.6222 0 15.1111C0 15.6 0.4 16 0.888889 16H15.1111C15.6 16 16 15.6 16 15.1111C16 14.6222 15.6 14.2222 15.1111 14.2222H14.2222V8.88889H15.1111ZM12.4444 14.2222H3.55556V8.88889C4.04445 8.88889 4.44445 8.48889 4.44445 8C4.44445 7.51111 4.04445 7.11111 3.55556 7.11111V1.77778H12.4444V7.11111C11.9556 7.11111 11.5556 7.51111 11.5556 8C11.5556 8.48889 11.9556 8.88889 12.4444 8.88889V14.2222Z"
+        fill="currentColor"
+      />
+    </svg>
+  );
+};
+
+export const RetailTradeIcon: FunctionComponent<IconProps> = ({ className }) => {
+  return (
+    <svg
+      width="19"
+      height="16"
+      viewBox="0 0 19 16"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      className={className}
+    >
+      <path
+        d="M18.3003 5.23556L17.367 1.35111C17.1714 0.551111 16.4781 0 15.6692 0H3.32255C2.52255 0 1.82033 0.56 1.63366 1.35111L0.700331 5.23556C0.486998 6.14222 0.682554 7.06667 1.25144 7.79556C1.32255 7.89333 1.42033 7.96445 1.50033 8.05334V14.2222C1.50033 15.2 2.30033 16 3.27811 16H15.7226C16.7003 16 17.5003 15.2 17.5003 14.2222V8.05334C17.5803 7.97333 17.6781 7.89333 17.7492 7.80445C18.3181 7.07556 18.5226 6.14222 18.3003 5.23556ZM15.6426 1.76889L16.5759 5.65333C16.6648 6.02667 16.5848 6.4 16.3537 6.69333C16.2292 6.85333 15.9626 7.11111 15.5181 7.11111C14.9759 7.11111 14.5048 6.67556 14.4426 6.09778L13.927 1.77778L15.6426 1.76889ZM10.3892 1.77778H12.1314L12.6114 5.79556C12.6559 6.14222 12.5492 6.48889 12.3181 6.74667C12.1226 6.97778 11.8381 7.11111 11.4737 7.11111C10.8781 7.11111 10.3892 6.58667 10.3892 5.94667V1.77778ZM6.38033 5.79556L6.86922 1.77778H8.61144V5.94667C8.61144 6.58667 8.12255 7.11111 7.46478 7.11111C7.16255 7.11111 6.887 6.97778 6.67366 6.74667C6.45144 6.48889 6.34478 6.14222 6.38033 5.79556ZM2.42478 5.65333L3.32255 1.77778H5.07366L4.55811 6.09778C4.487 6.67556 4.02478 7.11111 3.48255 7.11111C3.047 7.11111 2.77144 6.85333 2.65589 6.69333C2.41589 6.40889 2.33589 6.02667 2.42478 5.65333ZM3.27811 14.2222V8.86222C3.34922 8.87111 3.41144 8.88889 3.48255 8.88889C4.25589 8.88889 4.95811 8.56889 5.47366 8.04445C6.007 8.57778 6.71811 8.88889 7.527 8.88889C8.30033 8.88889 8.99366 8.56889 9.50922 8.06222C10.0337 8.56889 10.7448 8.88889 11.5448 8.88889C12.2914 8.88889 13.0026 8.57778 13.5359 8.04445C14.0514 8.56889 14.7537 8.88889 15.527 8.88889C15.5981 8.88889 15.6603 8.87111 15.7314 8.86222V14.2222H3.27811Z"
+        fill="currentColor"
+      />
+    </svg>
+  );
+};
+
+export const ExcelIcon: FunctionComponent<IconProps> = ({ className }) => {
+  return (
+    <svg
+      width="21"
+      height="20"
+      viewBox="0 0 21 20"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      className={className}
+    >
+      <path
+        d="M18.8125 11.6641V9.47656C18.8125 8.73064 18.5014 8.01527 17.9476 7.48782C17.3937 6.96038 16.6426 6.66406 15.8594 6.66406H14.5469C14.2858 6.66406 14.0354 6.56529 13.8508 6.38947C13.6662 6.21366 13.5625 5.9752 13.5625 5.72656V4.47656C13.5625 3.73064 13.2514 3.01527 12.6975 2.48782C12.1437 1.96038 11.3926 1.66406 10.6094 1.66406H8.96875M10.9375 1.66406H6.67188C6.1285 1.66406 5.6875 2.08406 5.6875 2.60156V16.9766C5.6875 17.4941 6.1285 17.9141 6.67188 17.9141H17.8281C18.3715 17.9141 18.8125 17.4941 18.8125 16.9766V9.16406C18.8125 7.17494 17.9828 5.26728 16.506 3.86076C15.0291 2.45424 13.0261 1.66406 10.9375 1.66406Z"
+        stroke="currentColor"
+        stroke-width="1.5"
+        stroke-linecap="round"
+        stroke-linejoin="round"
+      />
+      <path d="M10.5 9.78906L14 9.78906" stroke="currentColor" stroke-width="1.5" />
+      <path d="M14.875 9.78906H16.625" stroke="currentColor" stroke-width="1.5" />
+      <path d="M10.5 12.2891L14 12.2891" stroke="currentColor" stroke-width="1.5" />
+      <path d="M14.875 12.2891H16.625" stroke="currentColor" stroke-width="1.5" />
+      <path d="M10.5 14.7891L14 14.7891" stroke="currentColor" stroke-width="1.5" />
+      <path d="M14.875 14.7891H16.625" stroke="currentColor" stroke-width="1.5" />
+      <rect x="1.75" y="7.5" width="10.5" height="9.16667" rx="2" fill="#22A43F" />
+      <path
+        d="M5.80243 9.78906L6.97446 11.4932H7.01987L8.19757 9.78906H9.58527L7.81162 12.2891L9.625 14.7891H8.21176L7.01987 13.0825H6.97446L5.78257 14.7891H4.375L6.19405 12.2891L4.40905 9.78906H5.80243Z"
+        fill="white"
+      />
+    </svg>
+  );
+};
+
+export const PDFIcon: FunctionComponent<IconProps> = ({ className }) => {
+  return (
+    <svg
+      width="21"
+      height="20"
+      viewBox="0 0 21 20"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      className={className}
+    >
+      <path
+        d="M18.8125 11.6641V9.47656C18.8125 8.73064 18.5014 8.01527 17.9476 7.48782C17.3937 6.96038 16.6426 6.66406 15.8594 6.66406H14.5469C14.2858 6.66406 14.0354 6.56529 13.8508 6.38947C13.6662 6.21366 13.5625 5.9752 13.5625 5.72656V4.47656C13.5625 3.73064 13.2514 3.01527 12.6975 2.48782C12.1437 1.96038 11.3926 1.66406 10.6094 1.66406H8.96875M10.9375 1.66406H6.67188C6.1285 1.66406 5.6875 2.08406 5.6875 2.60156V16.9766C5.6875 17.4941 6.1285 17.9141 6.67188 17.9141H17.8281C18.3715 17.9141 18.8125 17.4941 18.8125 16.9766V9.16406C18.8125 7.17494 17.9828 5.26728 16.506 3.86076C15.0291 2.45424 13.0261 1.66406 10.9375 1.66406Z"
+        stroke="currentColor"
+        stroke-width="1.5"
+        stroke-linecap="round"
+        stroke-linejoin="round"
+      />
+      <rect x="1.75" y="8.33594" width="14.875" height="7.5" rx="2" fill="#C22424" />
+      <path
+        d="M3.5 14.1667V10H5.25803C5.59601 10 5.88393 10.0604 6.12182 10.1811C6.35971 10.3004 6.54102 10.4666 6.66577 10.6795C6.79196 10.8911 6.85506 11.1353 6.85506 11.4119C6.85506 11.6886 6.79124 11.9328 6.66359 12.1444C6.53594 12.356 6.351 12.5208 6.10877 12.6388C5.86798 12.7568 5.57642 12.8158 5.2341 12.8158H4.11357V12.1098H5.08179C5.26311 12.1098 5.41251 12.0806 5.53001 12.0223C5.64895 11.9626 5.73743 11.8806 5.79545 11.7761C5.85492 11.6703 5.88466 11.5489 5.88466 11.4119C5.88466 11.2736 5.85492 11.1529 5.79545 11.0498C5.73743 10.9454 5.64895 10.8647 5.53001 10.8077C5.41106 10.7494 5.26021 10.7202 5.07744 10.7202H4.44211V14.1667H3.5Z"
+        fill="white"
+      />
+      <path
+        d="M9.04662 14.1667H7.467V10H9.05967C9.50788 10 9.89372 10.0834 10.2172 10.2502C10.5407 10.4157 10.7894 10.6538 10.9635 10.9644C11.139 11.275 11.2268 11.6466 11.2268 12.0793C11.2268 12.5133 11.139 12.8863 10.9635 13.1982C10.7894 13.5102 10.5392 13.7496 10.2128 13.9164C9.88792 14.0833 9.49918 14.1667 9.04662 14.1667ZM8.40911 13.4119H9.00745C9.28595 13.4119 9.52021 13.3657 9.71023 13.2735C9.9017 13.1799 10.0453 13.0355 10.141 12.8402C10.2382 12.6435 10.2868 12.3899 10.2868 12.0793C10.2868 11.7714 10.2382 11.5198 10.141 11.3245C10.0453 11.1292 9.90243 10.9854 9.71241 10.8931C9.52239 10.8009 9.28813 10.7548 9.00963 10.7548H8.40911V13.4119Z"
+        fill="white"
+      />
+      <path
+        d="M11.9246 14.1667V10H14.875V10.7263H12.8668V11.7192H14.6792V12.4455H12.8668V14.1667H11.9246Z"
+        fill="white"
+      />
+    </svg>
+  );
+};
